+++ capstone_back/demo/src/main/java/com/example/demo/repository/ClientRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.entity.Client;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepository extends CrudRepository<Client, Integer> {//id타입 int
+public interface ClientRepository extends CrudRepository<Client, Long> {//id타입 int
 
 }
